Add 'is_url_valid'

+++ tests/testthat/test-install_tmhmm_bin.R
@@ -1,8 +1,7 @@
 test_that("use", {
-  skip("Cannot test installation if download link expires")
-
-  if (!curl::has_internet()) return()
   if (!is_on_travis()) return()
+  if (!curl::has_internet()) return()
+  if (!is_url_valid()) return()
 
   folder_name <- tempdir()
   if (dir.exists(folder_name)) unlink(folder_name, recursive = TRUE)
